react + vite

+++ backend/controllers/product.controller.js
@@ -1,20 +1,20 @@
-import Product from '../models/product.model.js';
-import mongoose from 'mongoose';
+import Product from "../models/product.model.js";
+import mongoose from "mongoose";
 
 export const getProducts = async (req, res) => {
   try {
     const products = await Product.find({});
     res.status(200).json({ sucess: true, data: products });
   } catch (error) {
-    console.log('Error in fetching products:', error);
+    console.log("Error in fetching products:", error);
     res.status(500).json({ success: false, message: "Server Error" });
   }
-}
+};
 
 export const createProduct = async (req, res) => {
   const product = req.body; //user will send the product data in the request body
   if (!product.name || !product.price || !product.image) {
-    return res.status(400).json({ message: 'Please enter all fields' });
+    return res.status(400).json({ message: "Please enter all fields" });
   }
 
   const newProduct = new Product(product);
@@ -22,11 +22,10 @@ export const createProduct = async (req, res) => {
   try {
     await newProduct.save();
     res.status(201).json({ success: true, data: newProduct });
-
   } catch (error) {
-    console.log('Error in creating product:', error);
+    console.log("Error in creating product:", error);
   }
-}
+};
 
 export const updateProduct = async (req, res) => {
   const { id } = req.params;
@@ -34,28 +33,37 @@ export const updateProduct = async (req, res) => {
   const product = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
   }
 
   try {
-    const updatedProduct = await Product.findByIdAndUpdate(id, product, { new: true });
+    const updatedProduct = await Product.findByIdAndUpdate(id, product, {
+      new: true,
+    });
     res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
     res.status(505).json({ success: false, message: "Server Error" });
   }
-}
+};
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   //console.log("id:", id);
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
-    res.status(200).json({ success: true, message: "Product deleted successfully" });
-
-
+    res
+      .status(200)
+      .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
-    console.log('Error in deleting product:', error);
-    res.status(404).json({ success: false, message: "Product not found" });
+    console.log("Error in deleting product:", error);
+    res.status(500).json({ success: false, message: "Server error" });
   }
-}
+};
